requirements.txt
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
binance-connector==2.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==39.0.2
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.0.0
requests==2.28.2
service-identity==21.1.0
six==1.16.0
SQLAlchemy==2.0.7
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
zope.interface==5.5.2
